Added schema parsing for building tables

+++ SilkRoute/Suppliers/suppliertable.h
@@ -13,8 +13,8 @@ public:
     {
            int id;
            QString name;
-           float profit;
-           float expenditure;
+           double profit;
+           double expenditure;
            int last_transaction;
     } SupplierData;
